+++ crates/holochain/src/core/workflow/error.rs
@@ -76,6 +76,9 @@ pub enum WorkflowError {
     DhtOpError(#[from] DhtOpError),
 
     #[error(transparent)]
+    DbCacheError(#[from] holochain_types::db_cache::error::DbCacheError),
+
+    #[error(transparent)]
     SysValidationError(#[from] SysValidationError),
 
     #[error(transparent)]
@@ -102,9 +105,6 @@ pub enum WorkflowError {
     /// Other
     #[error("Other: {0}")]
     Other(Box<dyn std::error::Error + Send + Sync>),
-
-    #[error("Tried to integrate activity {1} after {0} but agent activity must be in order")]
-    ActivityOutOfOrder(u32, u32),
 }
 
 impl WorkflowError {
