+++ crates/holochain_state/tests/cache_tests/mod.rs
@@ -1,18 +1,19 @@
 use arbitrary::Arbitrary;
 use arbitrary::Unstructured;
-use holo_hash::AnyDhtHash;
-use holo_hash::EntryHash;
-use holo_hash::HeaderHash;
+use holo_hash::*;
+use holochain_sqlite::prelude::*;
+use holochain_sqlite::rusqlite::Transaction;
 use holochain_state::validation_db::ValidationLimboStatus;
 use holochain_state::{mutations, prelude::test_in_mem_db};
-use holochain_types::dht_op::{DhtOpLight, OpOrder};
+use holochain_types::db_cache::*;
+use holochain_types::dht_op::{DhtOpLight, DhtOpType, OpOrder};
 use holochain_zome_types::Create;
 use holochain_zome_types::ValidationStatus;
-use holochain_zome_types::{Dna, Header, HeaderHashed, Signature, SignedHeaderHashed, NOISE};
-use rusqlite::Transaction;
-use test_case::test_case;
-
-use super::*;
+use holochain_zome_types::{
+    Dna, Header, HeaderHashed, Signature, SignedHeaderHashed, Timestamp, NOISE,
+};
+use std::collections::HashMap;
+use std::sync::Arc;
 
 fn insert_header_and_op(txn: &mut Transaction, u: &mut Unstructured, header: &Header) -> DhtOpHash {
     let timestamp = Timestamp::arbitrary(u).unwrap();
@@ -54,13 +55,9 @@ fn set_ready_to_integrate(db: &DbWrite<DbKindDht>, op_hash: &DhtOpHash) {
 
 async fn check_state(
     cache: &DhtDbQueryCache,
-    f: impl FnOnce(&HashMap<Arc<AgentPubKey>, ActivityBounds>),
+    f: impl FnOnce(&HashMap<Arc<AgentPubKey>, ActivityState>),
 ) {
-    cache
-        .get_or_try_init()
-        .await
-        .unwrap()
-        .share_ref(|activity| f(activity));
+    cache.get_state().await.share_ref(|activity| f(activity));
 }
 
 #[tokio::test(flavor = "multi_thread")]
@@ -77,12 +74,8 @@ async fn cache_inits_correctly() {
     let op_hash = db.test_commit(|txn| insert_header_and_op(txn, &mut u, &header));
 
     let cache = DhtDbQueryCache::new(db.clone().into());
-    check_state(&cache, |activity| {
-        let b = activity.get(header.author()).unwrap();
-        assert_eq!(b.integrated, None);
-        assert_eq!(b.ready_to_integrate, None);
-    })
-    .await;
+
+    check_state(&cache, |activity| assert!(activity.is_empty())).await;
 
     set_ready_to_integrate(&db, &op_hash);
 
@@ -257,12 +250,6 @@ async fn cache_set_integrated() {
     db.test_commit(|txn| insert_header_and_op(txn, &mut u, &header));
 
     let cache = DhtDbQueryCache::new(db.clone().into());
-    check_state(&cache, |activity| {
-        let b = activity.get(&author).unwrap();
-        assert_eq!(b.integrated, None);
-        assert_eq!(b.ready_to_integrate, None);
-    })
-    .await;
 
     cache
         .set_activity_ready_to_integrate(&author, 0)
@@ -276,9 +263,14 @@ async fn cache_set_integrated() {
     })
     .await;
 
+    check_state(&cache, |activity| {
+        dbg!(activity);
+    })
+    .await;
     cache.set_activity_to_integrated(&author, 0).await.unwrap();
 
     check_state(&cache, |activity| {
+        dbg!(activity);
         let b = activity.get(&author).unwrap();
         assert_eq!(b.integrated, Some(0));
         assert_eq!(b.ready_to_integrate, None);
@@ -361,163 +353,4 @@ async fn cache_set_all_integrated() {
         let range = test_activity.get(&author).unwrap();
         assert_eq!(*range, seq_range);
     }
-}
-
-#[test_case(1)]
-#[test_case(2)]
-#[test_case(u32::MAX)]
-#[test_case(u32::MAX - 1)]
-fn prev_is_empty_new_is_zero_check_empty(n: u32) {
-    let mut prev_bounds = ActivityBounds::default();
-    let mut new_bounds = ActivityBounds::default();
-    new_bounds.integrated = Some(n);
-    // () -> (n)
-    assert!(!prev_is_empty_new_is_zero(None, &new_bounds));
-    // (()) -> (n)
-    assert!(!prev_is_empty_new_is_zero(Some(&prev_bounds), &new_bounds));
-    prev_bounds.integrated = Some(5);
-    // (a) -> (n)
-    assert!(prev_is_empty_new_is_zero(Some(&prev_bounds), &new_bounds));
-}
-#[test]
-fn prev_is_empty_new_is_zero_check_zero() {
-    let prev_bounds = ActivityBounds::default();
-    let mut new_bounds = ActivityBounds::default();
-    new_bounds.integrated = Some(0);
-    // () -> (0)
-    assert!(prev_is_empty_new_is_zero(None, &new_bounds));
-    // (()) -> (0)
-    assert!(prev_is_empty_new_is_zero(Some(&prev_bounds), &new_bounds));
-}
-
-#[test_case(0)]
-#[test_case(1)]
-#[test_case(u32::MAX - 1)]
-fn integrated_is_consecutive_check_n(n: u32) {
-    let mut prev_bounds = ActivityBounds::default();
-    let mut new_bounds = ActivityBounds::default();
-    prev_bounds.integrated = Some(n);
-    new_bounds.integrated = Some(n + 1);
-    // (n) -> (n+1)
-    assert!(integrated_is_consecutive(Some(&prev_bounds), &new_bounds));
-    prev_bounds.integrated = None;
-    // () -> (n)
-    assert!(integrated_is_consecutive(None, &new_bounds));
-    // (()) -> (n)
-    assert!(integrated_is_consecutive(Some(&prev_bounds), &new_bounds));
-}
-
-#[test_case(0)]
-#[test_case(1)]
-#[test_case(u32::MAX)]
-fn integrated_is_consecutive_check_empty(n: u32) {
-    let mut prev_bounds = ActivityBounds::default();
-    let new_bounds = ActivityBounds::default();
-    // (n) -> ()
-    prev_bounds.integrated = Some(n);
-    assert!(integrated_is_consecutive(Some(&prev_bounds), &new_bounds));
-}
-
-#[test_case(0, 2)]
-#[test_case(0, u32::MAX)]
-#[test_case(1, 3)]
-#[test_case(1, u32::MAX)]
-#[test_case(u32::MAX - 2, u32::MAX)]
-#[test_case(u32::MAX, 0)]
-#[test_case(u32::MAX - 1, 0)]
-#[test_case(1, 0)]
-#[test_case(2, 0)]
-#[test_case(2, 1)]
-#[test_case(3, 1)]
-fn integrated_is_consecutive_check_finds_gaps(s: u32, e: u32) {
-    let mut prev_bounds = ActivityBounds::default();
-    let mut new_bounds = ActivityBounds::default();
-    // (s) -> (e)
-    prev_bounds.integrated = Some(s);
-    new_bounds.integrated = Some(e);
-    assert!(!integrated_is_consecutive(Some(&prev_bounds), &new_bounds));
-}
-
-#[test]
-fn can_accept_ready_in_random_order() {
-    use rand::prelude::*;
-    let mut activity = HashMap::new();
-    let mut u = Unstructured::new(&NOISE);
-    let mut rand = rand::thread_rng();
-    let test_activity: Vec<_> = std::iter::repeat_with(|| {
-        let mut seq: Vec<_> = (0..=50).collect();
-        seq.shuffle(&mut rand);
-        (AgentPubKey::arbitrary(&mut u).unwrap(), seq)
-    })
-    .take(100)
-    .collect();
-
-    let mut new_bounds = ActivityBounds::default();
-    for (author, sequence) in test_activity {
-        let mut spent = Vec::with_capacity(sequence.len());
-        for n in sequence {
-            spent.push(n);
-            spent.sort_unstable();
-            new_bounds.ready_to_integrate = Some(n);
-            update_activity(&mut activity, &author, &new_bounds).unwrap();
-            let current_top = spent
-                .iter()
-                .zip(spent.iter().skip(1))
-                .find(|(a, b)| **b != **a + 1)
-                .map(|(a, _)| *a)
-                .unwrap_or_else(|| *spent.last().unwrap());
-            assert_eq!(
-                activity.get(&author).unwrap().ready_to_integrate.unwrap(),
-                current_top
-            );
-        }
-    }
-}
-
-fn show_query(txn: &mut Transaction) {
-    let r: Vec<Result<(AgentPubKey, u32, Option<bool>, Option<bool>), _>> = txn
-        .prepare(
-            "
-SELECT
-  HA.author,
-  MIN(HA.seq) min_seq,
-  HA.hash IN (
-      SELECT * FROM (
-              SELECT
-                header_hash
-              FROM
-                DhtOp
-              WHERE
-                DhtOp.type = :register_activity
-                AND DhtOp.validation_stage = 3
-                AND DhtOp.validation_status IS NOT NULL
-      )
-  ),
-  HB.hash NOT IN (
-      SELECT * FROM (
-              SELECT
-                header_hash
-              FROM
-                DhtOp
-              WHERE
-                DhtOp.type = :register_activity
-                AND DhtOp.validation_stage = 3
-                AND DhtOp.validation_status IS NOT NULL
-      )
-  )
-FROM
-  Header AS HA
-  LEFT JOIN HEADER HB ON HB.prev_hash = HA.hash
-GROUP BY
-  HA.author
-        ",
-        )
-        .unwrap()
-        .query_map(
-            named_params! { ":register_activity": DhtOpType::RegisterAgentActivity},
-            |row| Ok((row.get(0)?, row.get(1)?, row.get(2)?, row.get(3)?)),
-        )
-        .unwrap()
-        .collect();
-    dbg!(r);
 }
