+++ crates/holochain_state/tests/integration.rs
@@ -1,0 +1,2 @@
+pub mod cache_tests;
+pub mod corrupt_db;
