+++ crates/holochain_types/Cargo.toml
@@ -36,6 +36,7 @@ mr_bundle = { path = "../mr_bundle", features = ["packing"], version = "0.0.6"}
 must_future = "0.1.1"
 nanoid = "0.3"
 observability = "0.1.3"
+parking_lot = "0.10"
 rand = "0.7"
 regex = "1.4"
 rusqlite = { version = "0.26" }
@@ -59,9 +60,11 @@ arbitrary = { version = "1.0", features = ["derive"], optional = true}
 contrafact = { version = "0.1.0-dev.1", optional = true }
 
 [dev-dependencies]
+arbitrary = "1.0"
 maplit = "1"
 matches = "0.1"
 serde_json = "1"
+test-case = "1.2.1"
 tokio = { version = "1.11", features = [ "full" ] }
 
 [features]
