+++ crates/holochain_types/src/lib.rs
@@ -19,6 +19,7 @@ pub mod autonomic;
 pub mod chain;
 pub mod combinators;
 pub mod db;
+pub mod db_cache;
 pub mod dht_op;
 pub mod dna;
 pub mod element;
@@ -31,6 +32,7 @@ pub mod link;
 mod macros;
 pub mod metadata;
 pub mod prelude;
+pub mod share;
 pub mod signal;
 pub mod validate;
 pub mod web_app;
