+++ crates/kitsune_p2p/kitsune_p2p/src/gossip/sharded_gossip.rs
@@ -35,7 +35,7 @@ use super::simple_bloom::{HowToConnect, MetaOpKey};
 
 pub use bandwidth::BandwidthThrottles;
 
-#[cfg(any(test, feature = "test_utils"))]
+#[cfg(feature = "test_utils")]
 #[allow(missing_docs)]
 pub mod test_utils;
 
@@ -50,7 +50,8 @@ mod store;
 mod bandwidth;
 mod next_target;
 
-#[cfg(any(test, feature = "test_utils"))]
+#[cfg(feature = "test_utils")]
+#[allow(missing_docs)]
 pub(crate) mod tests;
 
 /// max send buffer size (keep it under 16384 with a little room for overhead)
